requirements.txt
python-dotenv
spotipy
pyyaml